end of zoo with various _WIN32_WINNT versions
git-svn-id: http://svn.miranda-ng.org/main/trunk@13112 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/QuickContacts/src/quickcontacts.cpp
@@ -480,7 +480,7 @@ int CheckText(HWND hdlg, TCHAR *sztext, BOOL only_enable = FALSE)
 	if(sztext == NULL || sztext[0] == _T('\0'))
 		return 0;
 
-	int len = mir_tstrlen(sztext);
+	size_t len = mir_tstrlen(sztext);
 
 	if (only_enable)
 	{
